changed the app structure more clear and separated it to components
