+++ Todo list/src/types.ts
@@ -1,0 +1,8 @@
+import { Animated } from 'react-native';
+
+export interface Todo {
+  id: number;
+  title: string;
+  done: boolean;
+  opacity: Animated.Value;
+}
